tweak: CombatService
