+++ Code/client/World.cpp
@@ -18,6 +18,7 @@
 #include <Services/CalendarService.h>
 #include <Services/StringCacheService.h>
 #include <Services/PlayerService.h>
+#include <Services/CombatService.h>
 
 #include <Events/PreUpdateEvent.h>
 #include <Events/UpdateEvent.h>
@@ -46,6 +47,7 @@ World::World()
      ctx().emplace<CommandService>(*this, m_transport, m_dispatcher);
      ctx().emplace<PlayerService>(*this, m_dispatcher, m_transport);
      ctx().emplace<StringCacheService>(m_dispatcher);
+     ctx().emplace<CombatService>(*this, m_transport, m_dispatcher);
 }
 
 World::~World() = default;
