+++ Code/server/World.cpp
@@ -12,6 +12,7 @@
 #include <Services/OverlayService.h>
 #include <Services/CommandService.h>
 #include <Services/StringCacheService.h>
+#include <Services/CombatService.h>
 
 #include <es_loader/ESLoader.h>
 
@@ -34,6 +35,7 @@ World::World()
     ctx().emplace<OverlayService>(*this, m_dispatcher);
     ctx().emplace<CommandService>(*this, m_dispatcher);
     ctx().emplace<StringCacheService>(*this, m_dispatcher);
+    ctx().emplace<CombatService>(*this, m_dispatcher);
 
     ESLoader::ESLoader loader;
     // emplace loaded mods into modscomponent.
